+++ Data.lua
@@ -1,0 +1,20 @@
+local Name, Addon = ...
+
+Addon.dungeons = {
+    -- De Other Side
+    [29] = { start = {x = 425, y = -120, sublevel = 1} },
+    -- Halls of Atonement
+    [30] = { start = {x = 696, y = -303, sublevel = 1} },
+    -- Mists of Tirna Scithe
+    [31] = { start = {x = 768, y = -95, sublevel = 1} },
+    -- Plaguefall
+    [32] = { start = {x = 272, y = -115, sublevel = 1} },
+    -- Sanguine Depths
+    [33] = { start = {x = 100, y = -300, sublevel = 1} },
+    -- Spires of Ascension
+    [34] = { start = {x = 315, y = -395, sublevel = 1} },
+    -- The Necrotic Wake
+    [35] = { start = {x = 702, y = -230, sublevel = 1} },
+    -- Theater of Pain
+    [36] = { start = {x = 425, y = -442, sublevel = 1} },
+}